Move gather function execution into the Gatherer, also removes the link between gathering and insightsclient
