+++ pkg/controller/operator.go
@@ -137,7 +137,7 @@ func (s *Support) Run(ctx context.Context, controller *controllercmd.ControllerC
 	go periodic.Run(4, ctx.Done(), initialDelay)
 
 	authorizer := clusterauthorizer.New(configObserver)
-	insightsClient := insightsclient.New(nil, 0, "default", authorizer, clusterConfigGatherer)
+	insightsClient := insightsclient.New(nil, 0, "default", authorizer, gatherKubeConfig)
 
 	// upload results to the provided client - if no client is configured reporting
 	// is permanently disabled, but if a client does exist the server may still disable reporting
