+++ pkg/gather/clusterconfig/0_gahterer_test.go
@@ -1,4 +1,4 @@
-package record
+package clusterconfig
 
 import (
 	"reflect"
